chore: move quill to a component
