+++ src/client/routes/home/index.jsx
@@ -36,9 +36,7 @@ import {
     IconShieldLock,
 } from '@tabler/icons';
 import { useSelector } from 'react-redux';
-import ReactQuill from 'react-quill';
-import 'react-quill/dist/quill.bubble.css';
-
+import Quill from '../../components/quill';
 import QRLink from '../../components/qrlink';
 
 import { zipFiles } from '../../helpers/zip';
@@ -75,29 +73,6 @@ const Home = () => {
     const isLoggedIn = useSelector((state) => state.isLoggedIn);
 
     const { t } = useTranslation();
-
-    const modules = {
-        toolbar: [
-            [{ header: [1, 2, 3, 4, 5, 6, false] }],
-            ['bold', 'italic', 'underline', 'strike', 'blockquote'],
-            [{ list: 'ordered' }, { list: 'bullet' }, { indent: '-1' }, { indent: '+1' }],
-            ['link', 'image'],
-        ],
-    };
-
-    const formats = [
-        'header',
-        'bold',
-        'italic',
-        'underline',
-        'strike',
-        'blockquote',
-        'list',
-        'bullet',
-        'indent',
-        'link',
-        'image',
-    ];
 
     useEffect(() => {
         if (secretId) {
@@ -300,24 +275,12 @@ const Home = () => {
                         </Alert>
                     )}
 
-                    <ReactQuill
-                        theme="bubble"
+                    <Quill
                         defaultValue={t('home.maintxtarea')}
                         value={text}
                         onChange={onTextChange}
                         readOnly={inputReadOnly}
-                        modules={modules}
-                        formats={formats}
-                        preserveWhitespace
-                        style={{
-                            color: '#C1C2C5',
-                            backgroundColor: '#25262b',
-                            border: '0.0625rem solid #373A40',
-                            borderRadius: '0.25rem',
-                            fontSize: '0.875rem',
-                            zIndex: 1000,
-                            minHeight: secretId ? '5rem' : '13rem',
-                        }}
+                        secretId={secretId}
                     />
 
                     <Group grow>
